reorg configure
Split parsing code into more specific files

Get rid of unused code

Get rid of defaultCtx, just use an empty one insteads
